requirements.txt
Django==3.0.2
django-compressor==2.4
django-crispy-forms==1.8.1
django-sass-processor==0.8
libsass==0.19.4
